examples/crewai-python/pyproject.toml
[tool.poetry]
name = "crewai_e2b_python"
version = "0.1.0"
description = "Example project for showcasing E2B code interpreter with LangChain"
authors = ["Vasek Mlejnsky <vasek@e2b.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
e2b-code-interpreter = "^0.0.3"
langchain = "0.2.16"
python-dotenv = "^1.0.1"
langchain-groq = "^0.1.3"
langchainhub = "^0.1.15"
langchain-openai = "^0.1.3"
pydantic = "^2.9.2"

crewai = "^0.61.0"
crewai-tools = "^0.12.1"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "crewai_e2b_python.main:main"
